chore: move internal swiper logic in a vm hook
